Change the PID of echoed payload to the master PID
Currently when echoing a payload from a worker, the PID was left
untouched. This was resulting in the master recieving the echoed
payload to resend the payload to the master which echoed it.

With this commit, the PID of echoed payloads are updated to be the
PID of the master echoing the payload.